LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/css-module/content-type-checking-expected.txt

CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "css" is not valid

Harness Error (FAIL), message = TypeError: Import assertion type "css" is not valid

NOTRUN text/css
NOTRUN application/css
NOTRUN text/html+css
NOTRUN text/css;boundary=something
NOTRUN text/css;foo=bar
